Move base to stats
